The program is able to decompress and open the game
Now it is able to download the game, decompress it and open the .exe file inside the game folders.
Also the installation folder gets saved once you change the installation directory or on the first time you launch the application.
The GUI got a small design change, with a function that allows you to drag it around. Borders got removed, now it is ready to be customized even more deeply.